requirements.txt
fastapi==0.104.1
gunicorn==21.2.0
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pyyaml==6.0.1
python-dotenv==1.0.0
psutil==5.9.6
SQLAlchemy==1.4.50
PyMuPDF==1.23.8
pdf2image==1.16.3
pytesseract==0.3.10
Pillow==10.2.0
